feat: Move validationUtils inside validation folder + Fix lint error
